requirements.txt
MouseInfo==0.1.2
numpy==1.17.3
Pillow==6.2.1
Pmw==2.0.1
PyAutoGUI==0.9.48
PyGetWindow==0.0.7
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.25
PyTweening==1.0.3
